Untangle error types for embedded-hal 1.*.* impls
The error type specified by a SpiDevice does not need to be the same as
what is returned by the underlying SpiBus. The correct thing to do is to
get the error by treating the SpiDevice as a spi::ErrorType.

An unncessary type variable on the FDM SpiBus implementation was removed
by using spi::ErrorType::Error directly as well.